+++ sarvey/processing.py
@@ -504,8 +504,8 @@ class Processing:
             phase=unw_phase,
             num_points=point_obj.num_points,
             ifg_net_obj=point_obj.ifg_net_obj,
-            num_cores=1,  # self.config.general.num_cores,
             ref_idx=0,
+            num_cores=self.config.general.num_cores,
             logger=self.logger
         )
         point_obj = Points(file_path=join(self.path, "p1_ts.h5"), logger=self.logger)
@@ -575,8 +575,8 @@ class Processing:
         # for sbas the ifg network needs to be inverted to get the phase time series
         phase_ts = ut.invertIfgNetwork(phase=unw_phase, num_points=point_obj.num_points,
                                        ifg_net_obj=point_obj.ifg_net_obj,
-                                       num_cores=1,  # self.config.general.num_cores,
                                        ref_idx=0,
+                                       num_cores=self.config.general.num_cores,
                                        logger=self.logger)
 
         point_obj.phase = phase_ts
@@ -1071,7 +1071,7 @@ class Processing:
             phase=unw_phase,
             num_points=point2_obj.num_points,
             ifg_net_obj=point2_obj.ifg_net_obj,
-            num_cores=1,  # self.config.general.num_cores,
+            num_cores=self.config.general.num_cores,
             ref_idx=0,
             logger=self.logger)
 
@@ -1131,8 +1131,8 @@ class Processing:
 
         phase_ts = ut.invertIfgNetwork(phase=unw_phase, num_points=point_obj.num_points,
                                        ifg_net_obj=point_obj.ifg_net_obj,
-                                       num_cores=1,  # self.config.general.num_cores,
                                        ref_idx=0,
+                                       num_cores=self.config.general.num_cores,
                                        logger=self.logger)
 
         point_obj.phase = phase_ts